Extracted common profile functionality into base class.
